refacto  code ,  splash screen logo done splash screen animation done
