+++ ios/Runner/Base.lproj/LaunchScreen.storyboard
@@ -38,7 +38,7 @@
         </scene>
     </scenes>
     <resources>
-        <image name="LaunchImage" width="512" height="512"/>
+        <image name="LaunchImage" width="540" height="578"/>
         <image name="LaunchBackground" width="1" height="1"/>
     </resources>
 </document>
