+++ web/index.html
@@ -72,6 +72,10 @@
   
   
   
+  
+  
+  
+  
   <style id="splash-screen-style">
     html {
       height: 100%
@@ -144,6 +148,8 @@
       <source srcset="splash/img/dark-1x.png 1x, splash/img/dark-2x.png 2x, splash/img/dark-3x.png 3x, splash/img/dark-4x.png 4x" media="(prefers-color-scheme: dark)">
       <img class="center" aria-hidden="true" src="splash/img/light-1x.png" alt="">
   </picture>
+  
+
